+++ packages/shared/src/index.ts
@@ -7,3 +7,6 @@ export * from "./types";
 
 // Store
 export * from "./store/opsStore";
+
+// Utils
+export * from "./utils/buckets";